Create 2441. Largest Positive Integer That Exists With Its Negative---Leetcode
This repository contains a C++ solution for finding the maximum K from a given vector of integers. The findMaxK function utilizes sorting and a two-pointer technique to efficiently determine the maximum value. The code is well-structured and easy to understand, making it suitable for learning and implementation in similar scenarios. Feel free to explore, use, and contribute to this solution to enhance its functionality and performance.